api_modernization_tool/requirements.txt
# =============================================================================
# API MODERNIZATION TOOL - REQUIREMENTS
# =============================================================================
#
# Built for hackathons with free tier AWS resources
# Python 3.10+ required
#

# =============================================================================
# Core Python
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# Code Parsing (Tree-sitter) - OPTIONAL
# =============================================================================
# Uncomment if you want real Java parsing
# tree-sitter>=0.21.0
# tree-sitter-java>=0.21.0

# =============================================================================
# Graph Processing
# =============================================================================
networkx>=3.2

# =============================================================================
# LangChain Ecosystem (Latest Stable - 2025)
# =============================================================================
langchain>=0.3.13
langchain-core>=0.3.28
langchain-community>=0.3.13
langchain-text-splitters>=0.3.4

# =============================================================================
# Vector Store & Embeddings
# =============================================================================
faiss-cpu>=1.9.0

# Local embeddings (no API key needed, free!)
sentence-transformers>=3.3.1
torch>=2.0.0

# =============================================================================
# AWS Integration (Bedrock - Free tier available)
# =============================================================================
boto3>=1.28.0
botocore>=1.31.0

# =============================================================================
# Data Validation
# =============================================================================
pydantic>=2.10.0

# =============================================================================
# Web UI
# =============================================================================
streamlit>=1.41.0
streamlit-modal>=0.1.0

# =============================================================================
# Utilities
# =============================================================================
gitpython>=3.1.43
requests>=2.31.0
